LayoutTests/imported/w3c/web-platform-tests/html/user-activation/propagation-sameorigin.html

<!DOCTYPE html>
<html>
  <head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="../../resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
    <style>
      iframe {
        border: 1px solid red;
      }
    </style>
  </head>
  <body></body>
  <script>
    async function createNestedIframes() {
      const iframe = await attachIframe("resources/blank.html");
      const nested1 = await attachIframe(
        "resources/blank.html#nested1",
        iframe.contentDocument,
      );
      const nested2 = await attachIframe(
        "resources/blank.html#nested2",
        nested1.contentDocument,
      );
      return [iframe, nested1, nested2];
    }

    test(() => {
      assert_false(navigator.userActivation.isActive, "top isActive");
      assert_false(navigator.userActivation.hasBeenActive, "top hasBeenActive");
    }, "Initial user activation states");

    promise_test(async () => {
      const soloFrame = await attachIframe("resources/blank.html#solo");
      const nestedFrames = await createNestedIframes();
      const iframes = [soloFrame, ...nestedFrames];

      for (const frame of iframes) {
        const { userActivation } = frame.contentWindow.navigator;
        assert_false(
          userActivation.isActive,
          `before click isActive frame ${frame.src}`,
        );
        assert_false(
          userActivation.hasBeenActive,
          `before click hasBeenActive frame ${frame.src}`,
        );
      }

      await test_driver.click(soloFrame.contentDocument.body);

      for (const frame of iframes) {
        const { userActivation } = frame.contentWindow.navigator;
        assert_true(
          userActivation.isActive,
          `after click isActive frame ${frame.src}`,
        );
        assert_true(
          userActivation.hasBeenActive,
          `after click hasBeenActive frame ${frame.src}`,
        );
      }

      assert_true(navigator.userActivation.isActive, "top isActive");
      assert_true(navigator.userActivation.hasBeenActive, "top hasBeenActive");

      await consumeTransientActivation(soloFrame.contentWindow);

      assert_false(navigator.userActivation.isActive, "top isActive");
      assert_true(navigator.userActivation.hasBeenActive, "top hasBeenActive");
    }, "User activation propagates across same-origin frame boundary.");
  </script>
</html>
